# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/households/{householdId}/invitations": {
      "post": {
        "operationId": "InvitationsController_create",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invitation créée."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite un utilisateur à rejoindre le household par email (réservé à OWNER/ADMIN). En développement, le jeton d'acceptation est renvoyé en clair.",
        "tags": [
          "invitations"
        ]
      },
      "get": {
        "operationId": "InvitationsController_list",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitations en attente."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste les invitations en attente du household (réservé à OWNER/ADMIN).",
        "tags": [
          "invitations"
        ]
      }
    },
...
    "/households/{householdId}/items": {
      "get": {
        "operationId": "ItemsController_findAll",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "condition",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NEW",
                "VERY_GOOD",
                "GOOD",
                "FAIR",
                "POOR"
              ]
            }
          },
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "createdById",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "title",
                "createdAt",
                "updatedAt",
                "condition"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page de résultats paginée."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste les items du household, avec recherche, filtres, tri et pagination.",
        "tags": [
          "items"
        ]
      },
      "post": {
        "operationId": "ItemsController_create",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item créé."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ajoute un item au household, avec un ou plusieurs propriétaires.",
        "tags": [
          "items"
        ]
      }
    },
    "/households/{householdId}/items/{itemId}": {
      "get": {
        "operationId": "ItemsController_findOne",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item trouvé."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retourne le détail d'un item, avec ses métadonnées spécifiques.",
        "tags": [
          "items"
        ]
      },
      "patch": {
        "operationId": "ItemsController_update",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item mis à jour."
          },
          "400": {
            "description": "Les données envoyées sont invalides."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modifie un item (champs communs, métadonnées ou propriétaires).",
        "tags": [
          "items"
        ]
      },
      "delete": {
        "operationId": "ItemsController_archive",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item archivé."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive un item (suppression logique, restaurable).",
        "tags": [
          "items"
        ]
      }
    },
    "/households/{householdId}/items/{itemId}/restore": {
      "post": {
        "operationId": "ItemsController_restore",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Item restauré."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restaure un item précédemment archivé.",
        "tags": [
          "items"
        ]
      }
    },
...
    "/households/{householdId}/stats": {
      "get": {
        "operationId": "StatsController_getStats",
        "parameters": [
          {
            "name": "householdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistiques du household."
          },
          "401": {
            "description": "L'utilisateur n'est pas authentifié, ou le token est invalide ou expiré."
          },
          "403": {
            "description": "L'utilisateur n'a pas les droits nécessaires pour effectuer cette action."
          },
          "404": {
            "description": "La ressource demandée n'existe pas, ou n'appartient pas à ce household."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retourne les statistiques du household (total actif, par catégorie, par propriétaire, ajouts récents, archivés).",
        "tags": [
          "stats"
        ]
      }
    },
...
  },
  "info": {
    "title": "Notre Nid API",
    "description": "API de gestion de collection partagée pour Notre Nid.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateInvitationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "sam@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
...
      "BookMetadataDto": {
        "type": "object",
        "properties": {
          "author": {
            "type": "string"
          },
          "isbn": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "publicationYear": {
            "type": "number"
          },
          "language": {
            "type": "string"
          },
          "pageCount": {
            "type": "number"
          }
        }
      },
      "CdMetadataDto": {
        "type": "object",
        "properties": {
          "artist": {
            "type": "string"
          },
          "album": {
            "type": "string"
          },
          "releaseYear": {
            "type": "number"
          },
          "label": {
            "type": "string"
          },
          "format": {
            "type": "string"
          }
        }
      },
      "DvdMetadataDto": {
        "type": "object",
        "properties": {
          "director": {
            "type": "string"
          },
          "releaseYear": {
            "type": "number"
          },
          "edition": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "durationMinutes": {
            "type": "number"
          }
        }
      },
      "CreateItemDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "Dune"
          },
          "description": {
            "type": "string"
          },
          "condition": {
            "type": "string",
            "enum": [
              "NEW",
              "VERY_GOOD",
              "GOOD",
              "FAIR",
              "POOR"
            ]
          },
          "notes": {
            "type": "string"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "ownerIds": {
            "description": "IDs des membres propriétaires (au moins un)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "book": {
            "$ref": "#/components/schemas/BookMetadataDto"
          },
          "cd": {
            "$ref": "#/components/schemas/CdMetadataDto"
          },
          "dvd": {
            "$ref": "#/components/schemas/DvdMetadataDto"
          },
          "customMetadata": {
            "type": "object",
            "description": "Métadonnées pour une catégorie personnalisée"
          }
        },
        "required": [
          "categoryId",
          "title",
          "condition",
          "ownerIds"
        ]
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "Dune"
          },
          "description": {
            "type": "string"
          },
          "condition": {
            "type": "string",
            "enum": [
              "NEW",
              "VERY_GOOD",
              "GOOD",
              "FAIR",
              "POOR"
            ]
          },
          "notes": {
            "type": "string"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "ownerIds": {
            "description": "IDs des membres propriétaires (au moins un)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "book": {
            "$ref": "#/components/schemas/BookMetadataDto"
          },
          "cd": {
            "$ref": "#/components/schemas/CdMetadataDto"
          },
          "dvd": {
            "$ref": "#/components/schemas/DvdMetadataDto"
          },
          "customMetadata": {
            "type": "object",
            "description": "Métadonnées pour une catégorie personnalisée"
          }
        }
      }
    }
  }
}